refactor: padrão de nomes
