+++ src/domains/Artista/controllers/controllerArtista.ts
