+++ src/domains/Artista/services/serviceArtista.ts
